SOF-1539 Minor renaming
