+++ src/blueprints/tv2/factories/tv2-camera-factory.ts
@@ -4,7 +4,7 @@ import { ActionType } from '../../../model/enums/action-type'
 import { Tv2BlueprintConfiguration } from '../value-objects/tv2-blueprint-configuration'
 import { Tv2SourceMappingWithSound } from '../value-objects/tv2-studio-blueprint-configuration'
 import { TimelineObject } from '../../../model/entities/timeline-object'
-import { Tv2PieceMetadata } from '../value-objects/tv2-meta-data'
+import { Tv2PieceMetadata } from '../value-objects/tv2-metadata'
 import { PieceType } from '../../../model/enums/piece-type'
 import { Tv2AtemLayer, Tv2SisyfosLayer, Tv2SourceLayer } from '../value-objects/tv2-layers'
 import { PieceLifespan } from '../../../model/enums/piece-lifespan'
