+++ src/blueprints/tv2/tv2-end-state-for-part-service.ts
@@ -7,7 +7,7 @@ import { Tv2RundownPersistentState } from './value-objects/tv2-rundown-persisten
 import { Tv2TallyTags } from './value-objects/tv2-tally-tags'
 import { Tv2GraphicsContent } from './value-objects/tv2-content'
 import { Tv2SisyfosPersistentLayerFinder } from './helpers/tv2-sisyfos-persistent-layer-finder'
-import { Tv2SisyfosPersistenceMetadata } from './value-objects/tv2-meta-data'
+import { Tv2SisyfosPersistenceMetadata } from './value-objects/tv2-metadata'
 
 /*
  Disclaimer: The code in this file is almost a 1 to 1 copy of the code of the corresponding implementations in Blueprints.
